+++ Bootstrap/app/src/main/java/ar/com/wolox/android/example/di/AppModule.kt
@@ -5,13 +5,12 @@ import ar.com.wolox.android.example.ui.example.ExampleActivity
 import ar.com.wolox.android.example.ui.example.ExampleFragment
 import ar.com.wolox.android.example.ui.home.HomeActivity
 import ar.com.wolox.android.example.ui.home.HomeFragment
-import ar.com.wolox.android.example.ui.home2.Home2Activity
-import ar.com.wolox.android.example.ui.home2.Home2Fragment
+import ar.com.wolox.android.example.ui.home.NewsFragment
+import ar.com.wolox.android.example.ui.home.ProfileFragment
 import ar.com.wolox.android.example.ui.logIn.LogInActivity
 import ar.com.wolox.android.example.ui.logIn.LogInFragment
 import ar.com.wolox.android.example.ui.signUp.SignUpActivity
 import ar.com.wolox.android.example.ui.signUp.SignUpFragment
-
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -34,20 +33,22 @@ abstract class AppModule {
     internal abstract fun rootActivity(): RootActivity
 
     @ContributesAndroidInjector
-    internal abstract fun HomeFragment(): HomeFragment
-
-    @ContributesAndroidInjector
-    internal abstract fun HomeActivity(): HomeActivity
-
-    @ContributesAndroidInjector
     internal abstract fun SignUpActivity(): SignUpActivity
 
     @ContributesAndroidInjector
     internal abstract fun SignUpFragment(): SignUpFragment
 
     @ContributesAndroidInjector
-    internal abstract fun Home2Activity(): Home2Activity
+    internal abstract fun HomeActivity(): HomeActivity
 
     @ContributesAndroidInjector
-    internal abstract fun Home2Fragment(): Home2Fragment
+    internal abstract fun HomeFragment(): HomeFragment
+
+    @ContributesAndroidInjector
+    internal abstract fun NewsFragment(): NewsFragment
+
+    @ContributesAndroidInjector
+    internal abstract fun ProfileFragment(): ProfileFragment
+
+
 }
