+++ Bootstrap/app/src/main/java/ar/com/wolox/android/example/ui/home/HomeActivity.kt
@@ -1,13 +1,13 @@
-package ar.com.wolox.android.example.ui.home2
+package ar.com.wolox.android.example.ui.home
 
 import ar.com.wolox.android.R
 import ar.com.wolox.wolmo.core.activity.WolmoActivity
 
-class Home2Activity : WolmoActivity() {
+class HomeActivity : WolmoActivity() {
 
     override fun layout(): Int = R.layout.activity_base
 
     override fun init() {
-        replaceFragment(R.id.vActivityBaseContent, Home2Fragment())
+        replaceFragment(R.id.vActivityBaseContent, HomeFragment())
     }
 }
