--- Bootstrap/app/src/main/java/ar/com/wolox/android/example/ui/home/HomePresenter.java
@@ -1,16 +1,0 @@
-package ar.com.wolox.android.example.ui.home;
-
-import android.content.Context;
-import android.content.SharedPreferences;
-
-import javax.inject.Inject;
-
-import ar.com.wolox.wolmo.core.presenter.BasePresenter;
-
-class HomePresenter extends BasePresenter<IHomeView> {
-    @Inject
-    public HomePresenter(){
-
-    }
-}
-
