+++ Bootstrap/app/src/main/java/ar/com/wolox/android/example/ui/home/HomePresenter.kt
@@ -1,0 +1,9 @@
+package ar.com.wolox.android.example.ui.home
+
+
+import ar.com.wolox.wolmo.core.presenter.BasePresenter
+import javax.inject.Inject
+
+class HomePresenter @Inject constructor() : BasePresenter<IHomeView>() {
+
+}
