--- Bootstrap/app/src/main/java/ar/com/wolox/android/example/ui/home/IHomeView.java
@@ -1,4 +1,0 @@
-package ar.com.wolox.android.example.ui.home;
-
-interface IHomeView {
-}
