+++ Bootstrap/app/src/main/java/ar/com/wolox/android/example/ui/home/IHomeView.kt
@@ -1,0 +1,4 @@
+package ar.com.wolox.android.example.ui.home
+
+interface IHomeView {
+}
