--- Bootstrap/app/src/main/java/ar/com/wolox/android/example/ui/home2/Home2Presenter.kt
@@ -1,9 +1,0 @@
-package ar.com.wolox.android.example.ui.home2
-
-
-import ar.com.wolox.wolmo.core.presenter.BasePresenter
-import javax.inject.Inject
-
-class Home2Presenter @Inject constructor() : BasePresenter<IProfileView>() {
-
-}
