--- Bootstrap/app/src/main/java/ar/com/wolox/android/example/ui/home2/IHome2View.kt
@@ -1,4 +1,0 @@
-package ar.com.wolox.android.example.ui.home2
-
-interface IHome2View {
-}
