--- Bootstrap/app/src/main/res/layout/fragment_home2.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <android.support.v7.widget.Toolbar
-            android:id="@+id/fragment_home_topToolbar"
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/home_toolBar_layout_height"
-            android:background="@android:color/background_light"
-            android:minHeight="?attr/actionBarSize"
-            android:theme="?attr/actionBarTheme">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginStart="@dimen/static_spacing_medium_extra"
-                android:orientation="horizontal">
-
-                <ImageView
-                    android:id="@+id/fragment_home_wlogo"
-                    android:layout_width="@dimen/wlogo_layout_width"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/logo_marginStart"
-                    android:src="@drawable/wlogo" />
-
-                <TextView
-                    android:id="@+id/fragment_home_topToolbar_text"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/home_woloxText_marginStart"
-                    android:layout_marginTop="@dimen/home_woloxText_marginTop"
-                    android:layout_marginBottom="14dp"
-                    android:text="@string/wolox"
-                    android:textAlignment="viewStart"
-                    android:textAllCaps="false"
-                    android:textColor="@android:color/black"
-                    android:textSize="18sp"
-                    android:textStyle="bold"
-                    android:visibility="visible" />
-            </LinearLayout>
-
-        </android.support.v7.widget.Toolbar>
-
-        <android.support.v4.view.ViewPager
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/pager">
-
-
-            <android.support.design.widget.TabLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="top">
-
-                <android.support.design.widget.TabItem
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:clickable="false"
-                    android:icon="@drawable/news_tab_icon"
-                    android:text="@string/home_newsTab" />
-
-                <android.support.design.widget.TabItem
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:clickable="false"
-                    android:icon="@drawable/news_profile_tab"
-                    android:text="@string/home_profileTab" />
-            </android.support.design.widget.TabLayout>
-        </android.support.v4.view.ViewPager>
-    </LinearLayout>
-</FrameLayout>
