+++ Bootstrap/app/src/main/res/values/colors.xml
@@ -5,4 +5,6 @@
     <color name="colorPrimary">#a21d9bd7</color>
     <color name="colorPrimaryDark">#0960B7</color>
     <color name="colorAccent">#778dc63f</color>
+    <color name="white">#FFFFFF</color>
+    <color name="greenTraining">#8BB55A</color>
 </resources>
